refactor(entities): move entities module to maintainers domain
Restructure the codebase by relocating the entities module from the domains/entities directory to domains/maintainers.
This includes moving all related API, service, and type files, and updating import paths in dependent files.
