--- src/domains/entities/index.ts
@@ -1,3 +1,0 @@
-export * from './api';
-export * from './service';
-;
