+++ src/domains/maintainers/api/entitiesApi/entitiesApi.ts
@@ -1,6 +1,6 @@
-import { apiClient } from "../../../shared/services/apiService";
+import { apiClient } from "@/shared";
 import { ENTITIES_ENDPOINTS } from "./endpoints";
-import type { EntidadesApiResponse } from "../service/types";
+import type { EntidadesApiResponse } from "../../services/entitiesService";
 
 export const entitiesApi = {
     getClients: () => apiClient.get<EntidadesApiResponse>(ENTITIES_ENDPOINTS.GET_CLIENTS),
