+++ src/domains/maintainers/pages/clients/ClientListPage.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { UnderConstruction } from '@/components';
+
+export const ClientListPage: React.FC = () => {
+  return (
+    <UnderConstruction
+      title="Submodulo de clientes"
+      message="El submodulo de clientes está en construcción. Por favor, regresa pronto."
+    />
+  );
+};
