+++ src/domains/maintainers/services/index.ts
@@ -1,2 +1,3 @@
 export * from './MaintainersService';
 export * from './types';
+export * from './entitiesService';
